Revert "Feature/stnu"
