+++ src/temporal_networks/stnu/__init__.py
@@ -1,6 +1,6 @@
-__title__ = 'stn'
+__title__ = 'stntools'
 __version__ = '0.1.0'
 
 from .node import Node
 from .constraint import Constraint
-from .stn import STN
+from .stnu import STNU
