+++ src/temporal_networks/stnu/constraint.py
@@ -23,11 +23,10 @@
 # SOFTWARE.
 
 
-from temporal.networks.distempirical import norm_sample, uniform_sample
-from temporal.networks.stn import Constraint
+from src.temporal_networks.distempirical import norm_sample, uniform_sample
 
 
-class ConstraintSTNU(Constraint):
+class Constraint(object):
     """ Represents a temporal constraint between two nodes in the STN
         i: starting node
         j: ending node
@@ -47,14 +46,26 @@ class ConstraintSTNU(Constraint):
         i <--- -4 --- j
     """
 
-    def __init__(self, i=0, j=0, wji=-1, wij='inf', distribution=""):
-        super().__init__(i, j, wji, wij)
+    def __init__(self, i=0, j=0, wji=-1, wij=-1, distribution=None):
+        # node where the constraint starts
+        self.starting_node_id = i
+        # node where the constraint ends
+        self.ending_node_id = j
+        # Minimum allocated time between i and j
+        self.min_time = -wji
+        # Maximum allocated time between i and j
+        if wij == 'inf':
+            wij = float('inf')
+        self.max_time = wij
         # Probability distribution (for contingent constraints)
         self.distribution = distribution
         # Duration (for contingent constraints) sampled from the probability distribution
         self.sampled_duration = 0
         # The constraint is contingent if it has a probability distribution
-        self.is_contingent = distribution is not ""
+        self.is_contingent = distribution is not None
+
+    def __repr__(self):
+        return "Constraint {} => {} [{}, {}]".format(self.starting_node_id, self.ending_node_id, -self.min_time, self.max_time)
 
     def dtype(self):
         """Returns the distribution edge type as a String. If no there is
@@ -116,14 +127,13 @@ class ConstraintSTNU(Constraint):
         return float(name_split[1]) * 1000
 
     def to_dict(self):
-        constraint_dict = super().__to_dict()
-        # constraint_dict = dict()
-        # constraint_dict['starting_node_id'] = self.starting_node_id
-        # constraint_dict['ending_node_id'] = self.ending_node_id
-        # constraint_dict['min_time'] = - self.min_time
-        # if self.max_time == float('inf'):
-        #     self.max_time = 'inf'
-        # constraint_dict['max_time'] = self.max_time
+        constraint_dict = dict()
+        constraint_dict['starting_node_id'] = self.starting_node_id
+        constraint_dict['ending_node_id'] = self.ending_node_id
+        constraint_dict['min_time'] = - self.min_time
+        if self.max_time == float('inf'):
+            self.max_time = 'inf'
+        constraint_dict['max_time'] = self.max_time
         constraint_dict['distribution'] = self.distribution
         constraint_dict['sampled_duration'] = self.sampled_duration
         constraint_dict['is_contingent'] = self.is_contingent
@@ -131,25 +141,14 @@ class ConstraintSTNU(Constraint):
 
     @staticmethod
     def from_dict(constraint_dict):
-        constraint = Constraint.from_dict(constraint_dict)
-
-        starting_node_id = constraint.starting_node_id
-        ending_node_id = constraint.ending_node_id
-        min_time = - constraint.min_time
-        max_time = constraint.max_time
-        distribution = constraint_dict['distribution']
-        # sampled_duration = constraint_dict['sampled_duration']
-        # is_contingent = constraint_dict['is_contingent']
-
-        constraint_stnu = ConstraintSTNU(starting_node_id, ending_node_id, min_time, max_time, distribution)
-        # constraint = ConstraintSTNU()
-        # constraint.starting_node_id = constraint_dict['starting_node_id']
-        # constraint.ending_node_id = constraint_dict['ending_node_id']
-        # constraint.min_time = -constraint_dict['min_time']
-        # if constraint_dict['max_time'] == 'inf':
-        #     constraint_dict['max_time'] = float('inf')
-        # constraint.max_time = constraint_dict['max_time']
-        # constraint.distribution = constraint_dict['distribution']
-        # constraint.sampled_duration = constraint_dict['sampled_duration']
-        # constraint.is_contingent = constraint_dict['is_contingent']
-        return constraint_stnu
+        constraint = Constraint()
+        constraint.starting_node_id = constraint_dict['starting_node_id']
+        constraint.ending_node_id = constraint_dict['ending_node_id']
+        constraint.min_time = -constraint_dict['min_time']
+        if constraint_dict['max_time'] == 'inf':
+            constraint_dict['max_time'] = float('inf')
+        constraint.max_time = constraint_dict['max_time']
+        constraint.distribution = constraint_dict['distribution']
+        constraint.sampled_duration = constraint_dict['sampled_duration']
+        constraint.is_contingent = constraint_dict['is_contingent']
+        return constraint
