--- temporal/networks/stnu/__init__.py
@@ -1,6 +1,0 @@
-__title__ = 'stnu'
-__version__ = '0.1.0'
-
-from .node import NodeSTNU
-from .constraint import ConstraintSTNU
-from .stnu import STNU
